Moved more logic into shared package
