+++ packages/shared/src/parse.ts
@@ -1,4 +1,4 @@
-import { CustomElement, ErrorTypes } from './model';
+import { IProps, CustomElement, ErrorTypes } from './model';
 
 /* -----------------------------------
  *
@@ -51,8 +51,62 @@ function getDocument(html: string) {
 
 /* -----------------------------------
  *
+ * getAttributeObject
+ *
+ * -------------------------------- */
+
+function getAttributeObject(attributes: NamedNodeMap): IProps {
+  const result = {};
+
+  if (!attributes?.length) {
+    return result;
+  }
+
+  for (let i = attributes.length - 1; i >= 0; i--) {
+    const item = attributes[i];
+
+    result[item.name] = item.value;
+  }
+
+  return result;
+}
+
+/* -----------------------------------
+ *
+ * getAttributeProps
+ *
+ * -------------------------------- */
+
+function getAttributeProps(attributes: NamedNodeMap, allowed?: string[]): IProps {
+  const values = getAttributeObject(attributes);
+
+  let result = {};
+
+  for (const key of Object.keys(values)) {
+    if (allowed?.indexOf(key) === -1) {
+      continue;
+    }
+
+    result[getPropKey(key)] = values[key];
+  }
+
+  return result;
+}
+
+/* -----------------------------------
+ *
+ * Attribute
+ *
+ * -------------------------------- */
+
+function getPropKey(value: string) {
+  return value.replace(/-([a-z])/g, (value) => value[1].toUpperCase());
+}
+
+/* -----------------------------------
+ *
  * Export
  *
  * -------------------------------- */
 
-export { parseJson, getDocument };
+export { parseJson, getDocument, getPropKey, getAttributeObject, getAttributeProps };
